Moved conftest and updated spec
